[ot] target/riscv: fix ePMP access control to pmpaddr/pmpcfg
Fix several issues with ePMP checks on writes to pmpaddr/pmpcfg.

For pmpaddr: Rule Locking Bypass (mseccfg.RLB) was not implemented for writes
to locked entries/rules.

For pmpcfg: with Machine Mode Lockdown (mseccfg.MML) set (and RLB not set),
writes to pmpcfg would be allowed for the wrong cases of ePMP truth table.
The existing code restricts writes like so:
- L=1, X=1: cases 8, 10, 12, 14
- L=0, RWX!=WX: cases 0-2, 4-6
From the ePMP specification: "Adding a rule with executable privileges that
either is M-mode-only or a locked Shared-Region is not possible (...)"
This description matches cases 9-11, 13 of the truth table. This commit
implements the check by using pmp_get_epmp_operation to convert between PMP
configuration and ePMP truth table cases.

Signed-off-by: Loïc Lefort <loic@rivosinc.com>