[ot] hw/opentitan: ot_flash: Add prog fifo, watermarks and fifo fixes
This commit introduces the `prog_fifo` counterpart to the existing
`rd_fifo`, and adds additional functionality relating to the two FIFOs,
whilst fixing some existing issues. This includes properly resetting
related status and interrupt state values when FIFOs are reset. Basic
functionality for writing to the prog FIFO is added, though the
corresponding program operation is not yet implemented.

Watermarks have been added via the use of the FIFO_LVL register, and
these are hooked up to the corresponding interrupts.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>