[ot] hw/opentitan: ot_spi_host: rework command FIFO management
Observing Verilator waves, the command FIFO only contains next command,
not the one that is being executed. In other words, there can be 5 in-fly commands,
not 4, as 4 are stored in the command FIFO + 1 is being executed.

Remove CS channel limitation (from 8 to 32 bits) and command tracking value
(from 16 to 32 bits)

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>