[ot] hw/opentitan: ot_otp_eg: Stub out invariant def properties
Much more work needs to be done for Earlgrey's OTP functionality to
catch up to Darjeeling's OTP; this commit simply stubs out the
`inv_default_part_X` properties which are now being generated by
`cfggen.py`, despite only being present in Darjeeling's OTP and not in
Earlgrey's.

This allows newer config generations to be used with Earlgrey for QEMU
emulation.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>