[ot] hw/opentitan: ot_keymgr_dpe: Fix key / kmac app digest size check
The first assertion compares KMAC's key size to the KMAC's app digest
size, despite the comment suggesting that it should instead be the
(maximum) size of the keymgr_dpe's generated output key.

For the second assertion the two key sizes are equivalent, but compare
the keymgr_dpe key size rather than the kmac_key_size to more closely
align with the assertion message.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>