[ot] hw/opentitan: ot_keymgr: Add initial stubbed keymgr implementation
This adds a very basic stubbed keymgr implementation with all of the
different registers and fields (and some constant pkg values) defined.
This is entirely stubbed at the moment - all writes and reads to
registers are unimplemented, as are register reset values and any of
the internal logic. The intention is to build upon the keymgr
implementation in further commits.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>