[ot] scripts/opentitan: cfggen.py: Left-pad OTP constants to fill width
For Earlgrey's OTP, there is a constant value used that is one nibble
smaller than the full constant width. To avoid special casing in the
Earlgrey OTP controller, add additional logic to handle this in the
python scripts used by cfggen, which will automatically left-pad OTP
constants with 0s to match the width specified by the SystemVerilog
type.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>