[ot] hw/opentitan: ot_otp_dj: Fix incorrect loop bounds
When fetching entropy for ephemeral scrambling key generation, we
checked that `SRAM_NONCE_WORDS` were available but then iterate over
`SRAM_KEY_WORDS`. This does not cause problems because these are the
same size (4 words), but the correct definition should be used to
prevent potentially introducing issues in the future.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>