[ot] hw/opentitan: ot_flash: Refactor ot_flash_op_signal
Likely when this was first created it was intended to be used to signal
complete ops with different delays, but currently we use a different
`ot_flash_op_execute` and `ot_flash_op_complete` interface that performs
all operations synchronously where possible.

Since the `INIT` pseudo-operation / command is the only thing that uses
this, refactor the function to be a callback for `init_complete` after
some initialisation delay. Also add a note about refactoring the
misleading `OtFlashOperation` enum.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>