[ot] hw/opentitan: ot_flash: Fix info page cfg disable
Based on the data page configuration behaviour and the info page cfg
docs, I had previously assumed that `en=False` for an info page config
meant that no protection rules apply (and thus no errors). This is not
the case. From the RTL, the correct behaviour is instead to forbid any
access (like having all fields as false).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>