[ot] hw/opentitan: ot_flash: Centralize fifo status & interrupt updates
Refactor FIFO status updates to all occur after finishing the current
execution of an operation, rather than handling all the individual
status and interrupt logic.

This is equivalent and easier to understand/maintain, and will
importantly allow easier implementation of hardware flash controller
reads in the future. The flash controller and flash control arbiter are
designed such that these STATUS fifo fields and interrupts are only
triggered when idle or in a SW operation (i.e. when *not* in a HW
operation), since we don't want to get interrupts or see read fifo
status for a hardware read. This will be easier to disable for HW reads
later on now that it is entirely localised in one place.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>