[ot] hw/opentitan: ot_flash: Initial HW request support
Hardware flash requests are processed in the same way as flash
controller requests, but have key differences in how they update the
state of software-visible status registers and behave on errors (they
fault with a fatal error that needs a reset).

Hardware operation requests also have different protection rules
depending on the lifecycle phase, but that is stubbed as unimplemented
for now (all HW-sourced operations are allowed).

Hardware and software operations share a program FIFO, but do not share
read buffers. To facilitate sharing a FIFO, software is locked out of
writing to the program FIFO while there is not an active *sw-initiated*
program operation.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>