[ot] hw/opentitan: ot_flash: Add SW info page qualification checks
Some info pages that hold secrets (e.g. the creator and owner seeds for
key derivation by the keymgr, the isolated partition for provisioning,
etc.) are locked by an additional layer of protections on top of the
existing info page configuration registers, which depend on signals
broadcast by the lifecycle controller.

Since the lifecycle controller is not connected, these signals are
stubbed for now, but the logic for calculating the different permissions
on these pages and merging them with the existing sw register config is
all implemented to allow correct info page protection for software
operations.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>