[ot] hw/opentitan: ot_flash: Add HW request info page permissions
As with data pages, hardware requests bypass the sw-register-configured
protections and instead have a set of hardware protection rules applied
to them. The list of rules is searched for any matches (based on both
the specific info page and the current flash lcmgr phase), allowing
hardware only selective access to specific info pages containing secrets
during certain life cycle phases.

For these regions, the ECC and scramble properties can be optionally
disabled via the HW_INFO_CFG_OVERRIDE register to aid in recovery. This
is implemented as well. There are also OTP secrets that configure the
same properties, but these are left unimplemented for now.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>