[ot] hw/opentitan: ot_flash: Clear program FIFO on HW start / SW finish
As part of the logic for synchronizing the shared use of the program
FIFO between hardware and software flash operations (in addition to the
existing locking of SW writes to the PROG_FIFO), we always clear the
program FIFO at the start of a hardware operation, and clear any data
left over in the program FIFO at the end of a SW operation, regardless
of which operation it is (e.g. read, program, erase).

To be sure we only clear the program FIFO once for starting a HW
operation, create a new `ot_flash_op_start` function intended to be
called to start an operation, unlike `ot_flash_op_execute` which does
the bulk of the processing but can also be used to continue an operation
that did not previously complete processing. This also allowed nicely
removing some of the repeated tracing logic.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>