[ot] hw/opentitan: ot_keymgr: Implement KMAC KDF requests
This is mostly based on the equivalent keymgr_dpe implementation.

Adds a connection from the keymgr to the KMAC via the KMAC app
interface, and add logic for sending a KMAC request containing the
contents of the KDF buffer. Also stub out a function for handling
the KMAC response - the resulting logic will depend on the current
operation, but for now is stubbed.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>