[ot] hw/opentitan: ot_kmac: Handle re-entrant app requests
The KMAC app request API already errors if we have an existing request
pending from an APP, but there is an edge case with re-entrancy.

Consider:
1. A device calls a kmac->app_request(data).
2. The KMAC processes the data and responds, calling dev_app->fn(resp)
3. As a result of the response, the device wants to process more KMAC
   requests, so it sends another app request.
4. KMAC thinks the app is not pending but still sees the current app
   set, so it thinks it is receiving more data. It will try and process
   that data and then then transition to idle, but the data that was
   expected to be processed has not been processed.

This re-entrancy should be allowed because we are essentially just
trying to "queue" an operation now that the previous one is completed,
and we want it to run after the KMAC is idle again (i.e. state reset).
So, by simply modifying the existing check we can use the existing app
request/pending mechanisms to allow safe re-entrancy (also without
potential for stack issues).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>