[ot] hw/opentitan: ot_keymgr_dpe: Fix OP_STATUS.WIP mismatch after write
This is a quite subtle bug: software can write to the `OP_STATUS` and
`ERR_CODE` register fields to clear their values, but when there is an
ongoing operation, we would expect some of these status fields to be
continually asserted (e.g. `OP_STATUS.STATUS_WIP`). However, the main
FSM logic that would reassert these values is not invoked here.

More problematic is the fact that most keymgr_dpe operations will
take only a few FSM ticks, which means only a few FSM schedules with
the BottomHalf. The non-determinism introduced by this scheduling
means that you only quite rarely see this bug manifest (e.g. clear
`STATUS_WIP` and thus leave the `OP_STATUS` register with a value
of `STATUS_IDLE`, leading software to think that the keymgr_dpe is idle
(and thus an operation can be started), when in reality it is active).

To make sure the operation status updates accordingly, tick the FSM on
writes to this register also.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>