[ot] hw/opentitan: ot_keymgr: Add identity generation (output box) op
Implementing identity generation is a small addition on top of the
previously implemented Advance and Generate SW/HW output key operations.

The identity KDF input is just one of the constant seeds imported from
the configuration file as a property (depending on the current key stage
i.e. `keymgr`` FSM state). The key state is then sent to KMAC as normal,
and the output is as with a SW output command.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>