[ot] hw/opentitan: ot_kmac: Implement SW error handling
Previously, the KMAC `CMD.ERR_PROCESSED` bit was just clearing the
alerts, but not actually doing anything to the internal KMAC state.

This commit implements full error handling for both HW and SW operation
errors. The logic is the same for both types - upon an erroneous
command, the user must write the `err_processed` bit, upon which the
KMAC will flush its internal state by sending a `Process` command,
followed by an `Absorb` command when done. The key differences are:
 - For HW app interface errors: as soon as all the data is received, HW
   is responded to so we don't stall HW, but the FSM won't return to
   idle until SW acknowledges the error. If SW acknowledges the error
   before the HW request data is finished sending, then KMAC first waits
   for the full request, then responds, then returns to idle.
-  For SW app interface errors: the internal KMAC state is flushed as
   soon as the error is processed, ignoring any further message feed or
   commands and immediately processing, and then finishing ('Done')
   when absorbed, before returning to idle.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>