[ot] scripts/opentitan: tests: Mark rsa_4096_encryption_functest as flaky
This passes 99 out of 100 times for me, but had 1 failure with the
following error:

E00004 profile.c:15] CHECK-fail: cycles <= 4294967295U

As such, this seems to be timing related, and the test should be marked
as flaky on QEMU for now.

Long-term, the test should conditionally increase its max cycle bound in
the QEMU environment, as there is still value in running the test.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>