[ot] hw/opentitan: ot_i2c_host_proxy: I2C Host Proxy implementation
This device is a I2C host device that is externally driven over a
chardev interface. It can be used to issue read and write transfers with
devices on the QEMU I2C buses from outside.

This will primarily be used by `opentitanlib` for its QEMU "transport"
implementation, which will issue transfers primarily with the OpenTitan
I2C target block. However, all devices on the I2C bus are accessible,
and any driver that speaks the wire protocol can be used too.

This device is modelled as a I2C bus device, but it cannot be interacted
with by other bus devices and cannot have transactions targetting it.

Signed-off-by: Alice Ziuziakowska <a.ziuziakowska@lowrisc.org>