[ot] scripts/opentitan: tests: Add new i2c_qemu test
This test was merged in upstream `earlgrey_1.0.0` for the purpose of
some basic testing of the QEMU I2C flash device and so needs to be added
to the CI list here.

It turns out the test is still slightly flaky 10% of the time, so add it
as a flaky test for now until that can be figured out.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>