[ot] scripts/opentitan: regression: disable failing test
`sha256_functest` started failing when PR
https://github.com/lowRISC/opentitan/pull/28352
was merged on the OpenTitan side.

Our policy on QEMU is to disable the failing test if it was an OpenTitan
change that caused it to fail.

We should separately investigate why this test is now failing, though.
I wouldn't expect that particular pull request to change behaviour at
first glance.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>