[ot] hw/opentitan: ot_hmac: Update digest before reading
As workaround for an RTL issue, most SW will on partial writes wait for
a defined number of cycles and then save the digest context of a
completed message block without calling PROCESS/STOP. Currently the FIFO
is only processed/emptied when it becomes too full or one of these
commands is issued. Thus the current QEMU model cannot handle SW just
waiting for a block to be processed instead of directly issued
commands.

To emulate this correctly, make sure we process the FIFO and compute the
digest for the latest message block before any digest read. In most
cases the message FIFO will be empty and thus this will be (almost) a
no-op.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>