[ot] hw/opentitan: ot_keymgr: Only go idle on reset & op_status write
The keymgr should only change its `OP_STATUS` to idle upon reset, or if
it is explicitly cleared by SW. The `DONE_ERROR` and `DONE_SUCCESS`
operation status values need to be latched to be queried by software,
but subsequent FSM ticks (e.g. scheduled by the BH) can then change the
operation status back to idle.

In HW, this value is latched and only updated when `op_start`, i.e.
when the operation is ongoing. When the operation is not ongoing, the
value is thus not reset back to idle outside of SW writes/resets.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>