[ot] scripts/opentitan: flashgen.py: Change `otdesc` to use any binaries
Currently the `otdesc` option accepts an address for a binary and then
attempts to guess whether the binary is a ROM_EXT or a bootloader file,
loading it at the hard-coded offsets.

With the result of `opentitantool image assemble` it would instead be
more useful to place binaries arbitrarily, wherever we wish, even across
flash banks to support mirrored flash images. This commit changes the
`otdesc` option to just load the given data at the given offset provided
it does not exceed the maximum flash size. This option is basically
converted to mean "place this data at this location, without care for
what it is".

For now, use of ELFs for debug symbol support is dropped from the
`otdesc` option, because binaries can be loaded that are not ROM_EXTs or
bl0s as defined in the flash format. In the future, the flash format
should be updated to support an arbitrary number of convenience
debugging ELFs, and then this option can be added back, but support for
this is dropped for now when using this option (debug ELFs still remain
if loading ROM_EXT and bl0 files the normal way).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>