[ot] hw/opentitan: ot_vmapper: Fix range sort comparator logic
This commit fixes the comparator logic for sorting address ranges as
used by the vmapper device.

The current logic uses the difference between the `start` and `end`
addresses of each range, which are stored as `uint32_t` values (which is
correct, as virtual address ranges can cover any valid 32 bit address).
By subtracting the addresses in unsigned arithmetic we will underflow in
the case that the first address is smaller than the second.

Furthermore, the conversion to a signed `gint` (`int`) means that for
large address differences (e.g. consider a region starts at `0x00000000`
and another at 0x80000001 or higher) the result will when cast
overflow the 32-bit signed domain to become a negative value.

Replace this logic with comparisons that return simple unit sign
values. This now performs the intended sorting without risking overflow.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>